[ADD] logic to delete

+++ cron/clearLastDayNews.js
@@ -5,8 +5,11 @@ async function clearLastDayNews() {
     const currentDateTime = new Date();
     currentDateTime.setHours(currentDateTime.getHours() - 2);
     const iso8601Time = currentDateTime.toISOString();
-    const deleted=await News.deleteMany({ createdAt: { $lte: iso8601Time } });
-    return deleted.deletedCount;
+    const newsCount = await News.countDocuments();
+    if (newsCount > 100) {
+        const deleted = await News.deleteMany({ createdAt: { $lte: iso8601Time } });
+        return deleted.deletedCount;
+    }
 }
 
 export default clearLastDayNews;
